refactor: variants 이동
